simple-graph/requirements.txt
langchain
langchain-openai
pydantic
python-dotenv
langchain-tavily
langgraph
black
isort